Change Game.WIDTH/HEIGHT to Float

+++ core/src/se/battlegoo/battlegoose/gamestates/CreateLobbyState.kt
@@ -13,7 +13,7 @@ class CreateLobbyState : GameState() {
     protected var cam: OrthographicCamera = OrthographicCamera()
 
     init {
-        cam.setToOrtho(false, Game.WIDTH.toFloat(), Game.HEIGHT.toFloat())
+        cam.setToOrtho(false, Game.WIDTH, Game.HEIGHT)
     }
 
     private val background = Texture("placeholder.png")
